リファクタリング
・プログラムにおけるステータスワードの変数型を`Uint16`に統一
・ステータスワードを定数として扱うよう修正

+++ MaintenanceTool/WindowsExe/MaintenanceToolGUI/MaintenanceToolCommon/AppCommon.cs
@@ -243,7 +243,7 @@ namespace MaintenanceToolCommon
             return BitConverter.ToInt16(sub, 0);
         }
 
-        public static uint ToUInt16(byte[] value, int startIndex, bool changeEndian = false)
+        public static UInt16 ToUInt16(byte[] value, int startIndex, bool changeEndian = false)
         {
             byte[] sub = GetSubArray(value, startIndex, 2);
             if (changeEndian == true) {
